+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/util/LeaderCacheUtils.java
@@ -17,23 +17,18 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.connector.client;
+package org.apache.iotdb.db.pipe.connector.util;
 
 import org.apache.iotdb.common.rpc.thrift.TEndPoint;
 import org.apache.iotdb.common.rpc.thrift.TSStatus;
-import org.apache.iotdb.commons.exception.IllegalPathException;
-import org.apache.iotdb.commons.path.PartialPath;
 import org.apache.iotdb.rpc.TSStatusCode;
 
 import org.apache.tsfile.utils.Pair;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class LeaderCacheUtils {
-  private static final Logger LOGGER = LoggerFactory.getLogger(LeaderCacheUtils.class);
 
   private LeaderCacheUtils() {
     // Do nothing
@@ -45,35 +40,30 @@ public class LeaderCacheUtils {
    * @param status is the returned status after transferring a batch event.
    * @return a list of pairs, each pair contains a device path and its redirect endpoint.
    */
-  public static List<Pair<String, TEndPoint>> getRedirectionsAfterTransferBatch(TSStatus status) {
+  public static List<Pair<String, TEndPoint>> parseRecommendedRedirections(TSStatus status) {
     // If there is no exception, there should be 2 sub-statuses, one for InsertRowsStatement and one
     // for InsertMultiTabletsStatement (see IoTDBDataNodeReceiver#handleTransferTabletBatch).
-    List<Pair<String, TEndPoint>> redirectList = new ArrayList<>();
+    final List<Pair<String, TEndPoint>> redirectList = new ArrayList<>();
+
     if (status.getSubStatusSize() != 2) {
       return redirectList;
     }
 
-    for (TSStatus subStatus : status.getSubStatus()) {
+    for (final TSStatus subStatus : status.getSubStatus()) {
       if (subStatus.getCode() != TSStatusCode.REDIRECTION_RECOMMEND.getStatusCode()) {
         continue;
       }
 
-      for (TSStatus innerSubStatus : subStatus.getSubStatus()) {
+      for (final TSStatus innerSubStatus : subStatus.getSubStatus()) {
         if (innerSubStatus.isSetRedirectNode()) {
-          try {
-            new PartialPath(innerSubStatus.getMessage());
-            // The message field should be a device path.
-            redirectList.add(
-                new Pair<>(innerSubStatus.getMessage(), innerSubStatus.getRedirectNode()));
-          } catch (IllegalPathException e) {
-            LOGGER.warn(
-                "Illegal device path when updating leader cache: {}",
-                innerSubStatus.getMessage(),
-                e);
-          }
+          // We assume that innerSubStatus.getMessage() is a device path.
+          // The message field should be a device path.
+          redirectList.add(
+              new Pair<>(innerSubStatus.getMessage(), innerSubStatus.getRedirectNode()));
         }
       }
     }
+
     return redirectList;
   }
 }
